+++ library/src/main/java/com/anibalbastias/weatherlib/data/repository/WeatherRepositoryImpl.kt
@@ -1,0 +1,5 @@
+package com.anibalbastias.weatherlib.data.repository
+
+class WeatherRepositoryImpl {
+
+}