+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Implement Architecture for Demo and library
 * Some test for JFrog Artifactory
 * Clean code
+* implementation('com.anibalbastias.weatherapi:library:0.0.1-20200927.111159-1@aar')